backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Chamster <nikkovkel258@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
alembic = "^1.12.1"
python-dotenv = "0.15.0"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.2"}
fastapi-users-db-sqlalchemy = "^6.0.1"
asyncpg = "^0.29.0"
sqladmin = {extras = ["full"], version = "^0.16.0"}

[tool.poetry.group.dev.dependencies]
fastapi = "^0.104.1"
SQLAlchemy = "^2.0.23"
asyncpg = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"